# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Data Dagger API
  description: ""
  version: 0.1.9
...
paths:
...
  /api/auth/orgs:
    get:
      tags:
        - "Authentication | Organizations"
      summary: "Get all organizations for a logged-in user"
      description: "Return a list of all organizations for a logged-in user"
      responses:
        '200':
          description: "A list of invites"
        '401':
          description: "Unauthorized: User could not be authenticated."

    post:
      tags:
        - "Authentication | Organizations"
      summary: "Creates a new organization"
      description: "A new organization is created. Requires a valid JWT cookie on request. The user that makes the request will automatically get owner role of the organization"
      parameters:
        - in: body
          name: "Organization Info"
          schema:
            type: object
            required:
              - name
            properties:
              name:
                type: string
              contact_email:
                type: string
              contact_phone:
                type: string
      responses:
        '201':
          description: "Created: The new organization was successfully created."
        '400':
          description: "Bad Request: A name for the organization was not provided."
        '401':
          description: "Unauthorized: User could not be authenticated."
        '409':
          description: "Conflict: The organization name is unavailable."
...
  /api/auth/invite:
    get:
      tags:
        - "Authentication | Invites"
      summary: "Get logged-in user's invites"
      description: "Return a list of all invites for a logged-in user"
      responses:
        '200':
          description: "Ok"
        '401':
          description: "Unauthorized: User could not be authenticated."

    post:
      tags:
        - "Authentication | Invites"
      summary: "Invite users to an organization"
      description: "Endpoint for organization owners and moderators to invite new members"
      parameters:
        - in: body
          name: "request body"
          schema:
            type: object
            required:
              - org_id
              - emails
            properties:
              org_id:
                type: string
              emails:
                type: array
      responses:
        '200':
          description: "Ok"
          content:
            application/json:
              schema:
                type: object
                example:
                  invited: "An array of the successfully invited emails"
                  invalid: "An array of invalid email addresses"
                  alreadyJoined: "An array of emails by users that have already joined the organization"
                  tooEarly: "An array of emails that have active invitations sent less than 24 hours ago"
        '400':
          description: "Bad Request: The request includes illegal properties (not 'org_id' or 'emails'), 'org_id' or 'email' was not provided, properties are not of the correct type, or there are duplicate emails in the request"
        '401':
          description: "Unauthorized: User could not be authenticated."
        '403':
          description: "Forbidden: The user that made the request is not OWNER or MODERATOR in the organization."

    delete:
      tags:
        - "Authentication | Invites"
      summary: "Delete multiple invitations"
      description: "Endpoint for organization owners and moderators to delete pending invitations"
      parameters:
        - in: body
          name: "request body"
          schema:
            type: object
            required:
              - org_id
              - emails
            properties:
              org_id:
                type: string
              emails:
                type: object
      responses:
        '204':
          description: "No Content: The request was processed successfully"
        '400':
          description: "Bad Request: the request body did not include the correct fields or types."
        '401':
          description: "Unauthorized: User could not be authenticated."
        '403':
          description: "Forbidden: The user that made the request is not OWNER or MODERATOR in the organization."
  /api/auth/invite/{org_id}:
    get:
      tags:
        - "Authentication | Invites"
      summary: "Get invitations"
      description: "Returns all pending inviations for a given organization. (NOTE: only an org OWNER or MODERATOR can access this information)"
      parameters:
        - in: path
          name: org_id
          description: "the id for an organization"
      responses:
        '200':
          description: "A list of invites"
          content:
            application/json:
              schema:
                type: array
                example:
                  { data: [ { organization_id: "1234", email: "bob@johnson.com" }, { organization_id: "1234", email: "john@bobson.com" } ] }
        '401':
          description: "Unauthorized: User could not be authenticated."
        '403':
          description: "Forbidden: The user that made the request is not OWNER or MODERATOR in the organization."
  /api/auth/invite/answer:
    post:
      tags:
        - "Authentication | Invites"
      summary: "Answer invitations"
      description: "Deletes an invitation an adds the user to the organization if it was accepted"
      parameters:
        - in: body
          name: "request body"
          schema:
            type: object
            required:
              - org_id
              - answer
            properties:
              org_id:
                type: string
              answer:
                type: boolean
      responses:
        '204':
          description: "No Content: The request was processed successfully."
        '400':
          description: "Bad Request: the request body did not include the correct fields or types."
        '401':
          description: "Unauthorized: User could not be authenticated."
        '404':
          description: "Not Found: No user found with that id."

  # /admin
...